Added PDESkeleton class
Every PDE needs a finite element, a DoFHandler and a sparsity
pattern; moreover, these data can be shared among multiple distinct
PDE. The PDESkeleton class encapsulates all of this data.
The ShallowStream model no longer stores all of the fine-grained
data for scalar/vector problems; instead, it stores a scalar and a
vector PDESkeleton.
In future, this class will be reponsible for creating FEValues
objects rather than having all the boilerplate every time.
Additionally, it is the natural class to take responsibility for
parts of grid refinement.