Changed parameterization of regularization functionals
These used to take the smoothing length as an argument and store it as a
member of the class; instead, the smoothing length to some power is
multiplied by the regularization functional in the objective functional.
This way, we can run L-curve tests without recreating the regularization
functional object several times, which is unnecessary.