Added test for ice shelf derivative + hessian
I can't get this test to work if I check that the local quadratic
approximation to the action functional is accurate to 3rd order,
probably because of the way constraints are added to the Hessian matrix.
This patch changes the test to only look for 2nd-order approximation.
I should fix this later, somehow.