Revamp of plot module
This patch eliminates the catchall `plot` function in favor of several
functions that aim to roughly reproduce what matplotlib does. For
example, `tricontourf` takes in a finite element field or vector field
and plots filled contours; `streamplot` makes streamlines of a vector
field; and so forth. Most importantly, the return types for these
functions are the same as their counterparts in matplotlib, making it
easier to add things like colorbars.