Getting rid of inequality constraints in inverse by reparameterizing
Enforcing the fact that parameters like the ice fluidity or friction
coefficient must be positive in the inverse solve by adding inequality
constraints and barrier functions introduces lots of hard-to-debug
complications. This patch changes the implementation to assume that all
parameter values are allowed by reparameterizing the fluidity in terms
of a new field using e.g. the exponential function.