Switching to CircleCI for testing
The test suite is taking long enough to run up against the 50-min limit
that the free tier of Travis CI offers. Moreover, a private equity firm
bought the company that maintains Travis and proceeded to fire most of
the engineering staff. This patch adds a config for the competing
service CircleCI and removes the old config for Travis.