Using pooch to fetch all data
This patch uses the library [pooch](https://github.com/fatiando/pooch)
to fetch all of the data and meshes instead of the icepack-data repo.
This eliminates the need for users to have the icepack-data repo on their
machines at all to run the demos. That repository has been repurposed as
as a place to keep glacier outlines and scripts for making those
outlines easy to mesh. The script `mesh.py` in the demos folder will
turn these outlines into a mesh using
[pygmsh](https://github.com/nschloe/pygmsh). Both pooch and pygmsh have
been added as dependencies.