Revert back to bespoke optimization solver
Revert commits:
* 6167dd02df7094b70d668a45762fc35cd342c28d
* ac37ab73521e73545982110782405cf04729d39b
* e4d9c3b4d4e4b30b4e2254d56204539f6b076f52
* ce922849a71ebf3f09a3854a08213b719104ac27
* f8ed459946b1f4245c3981dfe1613c037c5d33af

Transitioning to PETSc SNES was too hasty; the hand-written optimization
solver appears to be more robust to the kind of pathological input data
that one encounters at intermediate stages of an inverse solver. It
might be feasible to switch to PETSc SNES at a later date but this will
require either trust region methods or bounds-constrained optimization
to keep the parameters guesses sensible.