Fix bug in flow solver options setting
The flow solver setup routines were using the `.pop` method on
dictionaries other than `**kwargs`. The kwargs dict is special because
it gets copied, so you can modify it and code outside that block won't
see the modifications. This was causing a subtle bug in the how-to guide
on perf tuning.