1.5-Dimensional Flowline Model (#65)
* Initial commit for flowline model

Starting to add some functionality for an optional 1.5-dimensional
model. This is built into the hybrid model and will give an option for
flowline modeling. 1.5 dimensional is x and 'partial' z dimensions.

* Toward 1.5-d compatability

Checking here that 2.5-d hybrid model still works. Had to make some
small changes to clean it up. Those, along with some more changes toward
1.5-d.

* 1.5-d model working

Initial tests of diagnostic and prognostic models are good.

* Add a test for the flowline model

Nearly identical to the 2.5d hybrid model test.

* Some tests to make sure dimension choice tree is working

- flowline_test is mostly working but there is an incompatibility between
the 1-d array and 1.5-d array. This should be the same for 2-d and 2.5-d
but that is how it is written in hybrid_test.

- The ice stream demo is working in both 1-d and 2-d.

- The hybrid demo is working in both 1.5-d and 2.5-d.

- It seems like the 2-d and 2.5-d demos are slower than before?

* Clean up by removing dimensions variable

Tracking the dimension variable thoughtout the model class was unneccessary. The
dimension is now called when it is needed.

* Change dimension nomenclature to strings instead of float

This commit takes care of most issues in the pull request. Many were
small, but the model dimension is now named by a string (i.e. 'x', 'xy',
etc.).

* Update flow solver for facet_normal_nd function.

* Parameterize the hybrid test to include the flowline case ('xz')

* Update function ds_nd to be compatible with string dimension, 'xz' etc.

* Final changes to cleanup before merging into master.

* Fix import goof

Co-authored-by: Daniel Shapero <shapero.daniel@gmail.com>