Test data gridding
Note that this exposed a bug if the mesh bounds are right at an x or y
value of the raster grid, but with a rounding error. Floor of the
row_off with window in rasterio can go the wrong way in this case
(maybe only for y, which can be flipped), so padding by 2 is not
sufficient when constructing the bounds. This is fixed by padding by 3
(and using the yres for y), but may still be a problem if the mesh is
truly at the edge of the raster.