Bug fix in Triangle mesh translation
We need to add `reorder=False` when creating the DMPlex in order to
guarantee that the cells of the Firedrake and Triangle meshes are
numbered the same way. The numbering has to be the same if we want to
use some expression on the Firedrake side to determine how to refine the
Triangle mesh.